Memomize calls to MB within each run.
The vast majority of the program time is spent calling MusicBrainz
or waiting before making the next call to MusicBrainz. I realized
that by remembering the results of previous queries we might be
able to avoid some of these calls.

The results were better than I expected. Between 50% and 70% of
queries to MusicBrainz now hit the cache and do not have to perform
the query again. Because this MusicBrainz was nearly all of the
program's run time, this means that the program now runs in 30% to 50%
of that time that it used it.

In cache.py, I also have some code for an even more aggressive,
experimental memoization scheme in which queries are remembered
across runs. I'm not even sure if this is a good idea, but it could
yield huge speed-ups for some cases. Unfortunately, my implementation,
which read pickled Python objects at start-up and wrote them to the
disk before shutting down, took so long as to outweigh its benefits.
For this to work, the cache results we need to stay loaded in memory
between runs and perhaps ought to use something more structured than
a Python dictionary for storage.