More work to support non-ASCII characters.
I added the toUnicode function which takes an object or a string
or a unicode and intelligently returns a unicode.

This function should called on any strings that come from the
outside world to ensure that we always use unicode within the program.

Hopefully this will finally resolve the string encoding issues.