Caching results across runs using Sqlite3 DB.
I think (though I'm not 100% sure) that is really only useful
for developers who want to rerun Audiolog on the same inputs
multiple times. Nonetheless, it's simple, lightweight and allows
us to do very fast (~1s/sample) testing without the hassle of
running a local instance of the MusicBrainz server.